+++ CHANGELOG.md
@@ -1,5 +1,7 @@
 # Unreleased
 
+* Adding common packer templates
+
 # 1.1.0
 
 * Common Makefile commands